Refactor server into a shared fixture
Refactor the HTTP server used by the tests into a shared fixture.
This allows tests to be run in parallel again and generally improves
test throughput.
