+++ tests/Website.Tests/Integration/HttpServerCollection.cs
@@ -1,0 +1,19 @@
+﻿// Copyright (c) Martin Costello, 2016. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
+
+namespace MartinCostello.Website.Integration
+{
+    using Xunit;
+
+    /// <summary>
+    /// A class representing the collection fixture for an HTTP server.
+    /// </summary>
+    [CollectionDefinition(Name)]
+    public sealed class HttpServerCollection : ICollectionFixture<HttpServerFixture>
+    {
+        /// <summary>
+        /// The name of the test fixture.
+        /// </summary>
+        public const string Name = "HTTP server collection";
+    }
+}
