+++ tests/Website.Tests/Integration/Pages/HomePageTests.cs
@@ -8,6 +8,15 @@ namespace MartinCostello.Website.Integration.Pages
     /// </summary>
     public class HomePageTests : HtmlPageTest
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="HomePageTests"/> class.
+        /// </summary>
+        /// <param name="fixture">The fixture to use.</param>
+        public HomePageTests(HttpServerFixture fixture)
+            : base(fixture)
+        {
+        }
+
         /// <inheritdoc/>
         protected override string Path => "/";
     }
