+++ tests/Website.Tests/Integration/Pages/ProjectsPageTests.cs
@@ -8,6 +8,15 @@ namespace MartinCostello.Website.Integration.Pages
     /// </summary>
     public class ProjectsPageTests : HtmlPageTest
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ProjectsPageTests"/> class.
+        /// </summary>
+        /// <param name="fixture">The fixture to use.</param>
+        public ProjectsPageTests(HttpServerFixture fixture)
+            : base(fixture)
+        {
+        }
+
         /// <inheritdoc/>
         protected override string Path => "/projects/";
     }
