+++ tests/Website.Tests/xunit.runner.json
@@ -1,4 +1,3 @@
 {
-  "methodDisplay": "method",
-  "parallelizeTestCollections":  false
+  "methodDisplay": "method"
 }
